Refactor chatHistory state to Jotai
